Methods

This page documents the generic confint and pvalue methods which are supported by most tests. Some particular tests support additional arguments: see the documentation for the relevant methods provided in sections covering these tests.

Confidence interval

StatsAPI.confint — Function

confint(test::HypothesisTest; level = 0.95, tail = :both)

Compute a confidence interval C with coverage level.

If tail is :both (default), then a two-sided confidence interval is returned. If tail is :left or :right, then a one-sided confidence interval is returned.

Note

Most of the implemented confidence intervals are strongly consistent, that is, the confidence interval with coverage level does not contain the test statistic under $h_0$ if and only if the corresponding test rejects the null hypothesis $h_0: θ = θ_0$:

\[ C (x, level) = \{θ : p_θ (x) > 1 - level\},\]

where $p_θ$ is the pvalue of the corresponding test.

p-value

HypothesisTests.pvalue — Function

pvalue(test::HypothesisTest; tail = :both)

Compute the p-value for a given significance test.

If tail is :both (default), then the p-value for the two-sided test is returned. If tail is :left or :right, then a one-sided test is performed.